pyproject.toml
[tool.poetry]
name = "SanriBackend"
version = "0.1.0"
description = ""
authors = ["Ulukman Amangeldiev <ulukmanatageldiuulu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
drf-yasg = "^1.20.0"
djangorestframework = "^3.12.4"
Django = "^3.2.4"
Pillow = "^8.2.0"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
isort = "^5.8.0"
flake8 = "^3.9.2"
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = 'env/'
line_length = 79

[tool.isort]
skip_gitignore = true

[tool.autoflake]
remove_unused_variables = true
remove_all_unused_imports = true